+++ src/NAR.h
@@ -38,7 +38,7 @@
 //Parameters//
 //----------//
 #define NAR_DEFAULT_TRUTH ((Truth) { .frequency = NAR_DEFAULT_FREQUENCY, .confidence = NAR_DEFAULT_CONFIDENCE })
-extern long currentTime;
+extern uint32_t currentTime;
 
 //Callback function types//
 //-----------------------//
