+++ src/Stats.c
@@ -24,17 +24,17 @@
 
 #include "Stats.h"
 
-long Stats_countConceptsMatchedTotal = 0;
-long Stats_countConceptsMatchedMax = 0;
+uint32_t Stats_countConceptsMatchedTotal = 0;
+uint32_t Stats_countConceptsMatchedMax = 0;
 
-void Stats_Print(long currentTime)
+void Stats_Print(uint32_t currentTime)
 {
     puts("Statistics");
-    printf("countConceptsMatchedTotal:\t%ld\n", Stats_countConceptsMatchedTotal);
-    printf("countConceptsMatchedMax:\t%ld\n", Stats_countConceptsMatchedMax);
-    long countConceptsMatchedAverage = Stats_countConceptsMatchedTotal / currentTime;
-    printf("countConceptsMatchedAverage:\t%ld\n", countConceptsMatchedAverage);
-    printf("currentTime:\t\t\t%ld\n", currentTime);
+    printf("countConceptsMatchedTotal:\t%" PRIu32 "\n", Stats_countConceptsMatchedTotal);
+    printf("countConceptsMatchedMax:\t%" PRIu32 "\n", Stats_countConceptsMatchedMax);
+    uint32_t countConceptsMatchedAverage = Stats_countConceptsMatchedTotal / currentTime;
+    printf("countConceptsMatchedAverage:\t%" PRIu32 "\n", countConceptsMatchedAverage);
+    printf("currentTime:\t\t\t%" PRIu32 "\n", currentTime);
     printf("total concepts:\t\t\t%d\n", concepts.itemsAmount);
     printf("Maximum chain length in concept hashtable: %d\n", HashTable_MaximumChainLength(&HTconcepts));
     printf("Maximum chain length in atoms hashtable: %d\n", HashTable_MaximumChainLength(&HTatoms));
