+++ src/Usage.c
@@ -24,14 +24,14 @@
 
 #include "Usage.h"
 
-double Usage_usefulness(Usage usage, long currentTime)
+double Usage_usefulness(Usage usage, uint32_t currentTime)
 {
     double recency = currentTime - usage.lastUsed;
     double usefulnessToNormalize = ((double) usage.useCount) / (recency + 1.0);
     return usefulnessToNormalize / (usefulnessToNormalize + 1.0);
 }
 
-Usage Usage_use(Usage usage, long currentTime, bool eternalInput)
+Usage Usage_use(Usage usage, uint32_t currentTime, bool eternalInput)
 {
     return (Usage) { .useCount = usage.useCount+(eternalInput ? ETERNAL_INPUT_USAGE_BOOST : 1),
                      .lastUsed = currentTime };
@@ -39,5 +39,5 @@ Usage Usage_use(Usage usage, long currentTime, bool eternalInput)
 
 void Usage_Print(Usage *usage)
 {
-    printf("Usage: useCount=%ld lastUsed=%ld\n", usage->useCount, usage->lastUsed);
+    printf("Usage: useCount=%" PRIu32 " lastUsed=%" PRIu32 "\n", usage->useCount, usage->lastUsed);
 }
